refactor: rename page folders and update their imports in routes.ts file
