+++ client/src/app/router/routes.ts
@@ -1,8 +1,8 @@
 import type { RouteRecordRaw } from "vue-router";
-import Landing from "@/pages/landingPage/index.vue";
-import Login from "@/pages/loginPage/index.vue";
-import Signup from "@/pages/signupPage/index.vue";
-import ChatPage from "@/pages/chatPage/index.vue";
+import Landing from "@/pages/landing/index.vue";
+import Login from "@/pages/login/index.vue";
+import Signup from "@/pages/signup/index.vue";
+import ChatPage from "@/pages/chat/index.vue";
 
 export const routes: RouteRecordRaw[] = [
     {
