+++ client/src/pages/landing/index.vue
